Add files via upload
This is my design document that will be updated as I work on it further. I got this template from a friend, and after a course on Operating Systems, I learned that technical specs are really important, especially for any project that would take more than one programmer or a week of effort. Considering I am unfamiliar with physics simulations, this project will likely take a few weeks, so this is my current design document I am filling out. Not finished yet.